refactor: update DomainException import paths to use dot notation
