+++ src/domain/exceptions/domain.exception.ts
